Implementación de ZoneController

+++ src/main/java/com/nautik/api/domain/moorings/MooringDimension.java
@@ -27,7 +27,7 @@ public class MooringDimension {
 
     @NotNull
     @Column(name = "max_draft", nullable = false)
-    private Long maxDraft;
+    private Integer maxDraft;
